modify sample query panel
