+++ lib/utils.js
@@ -53,7 +53,7 @@ const CQT_NUM_QUERIES              = 5;
 const ALL_TERM_ITEM = 'All Terms';   // used to indicate no specific item is selected
 const NO_TERM_ITEM = 'Select Term';  // used to indicate no specific item is selected
 const NO_PRODUCT_ITEM = 'Select Product';       // used to indicate no specific product is selected
-const NO_CATEGORY_SELECTED = 'Select Category'; // used to indicate no specific category is selected
+const NO_CATEGORY_SELECTED = ' '; // used to indicate no specific category is selected
 
 // filter types and strings to use
 const filterTypes = [
